clicking on a favorite queue button when in a lobby of that queue already doesnt make a new lobby anymore

+++ src/FavoriteQueues/index.js
@@ -22,6 +22,8 @@ function updateFavoriteMenu() {
     queueElem.textContent = queueInfo["description"]
     queueElem.addEventListener("click", async (event) => {
       event.stopPropagation();
+      const lobbyInfo = await fetch("/lol-lobby/v2/lobby", {method: "GET"}).then(async r => await r.json())
+      if (lobbyInfo["gameConfig"]["queueId"] == queueId) return
       await fetch("/lol-lobby/v2/lobby", {method: "DELETE"})
       await fetch("/lol-lobby/v2/lobby", {
         method: "POST",
